move DownloadState to new file
